Adds more garbage collection.
Fixes #6. #6 was caused not by nitrogen disappearing, but by the turfs above the vents no longer sharing--this was due to them becoming "invalid" due to their constantly lowering amounts of certain trace gases ending up with denormals in their moles list, causing them to be considered "corrupt", causing FDA to ignore them. This is fixed by garbage collection, so this bug goes with it.